third_party/WebKit/LayoutTests/inspector/timeline/timeline-layer-tree-snapshot-expected.txt

Tests the Timeline API instrumentation of a Layout event

Layer tree dump:
#document
    html > body > div:nth-child(2) 100x100
Layer tree dump:
#document
    html > body > div:nth-child(2) 100x100
    html > body > div:nth-child(3) 100x100
